Completely remove the WAABBHierarchy structure.
It is now replaced by the WQuadtree.
